requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
bcrypt==4.1.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-redis==5.4.0
idna==3.6
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
